Conflicts:
	public/assets/css/style.css
	resources/views/auth/register.blade.php
	routes/web.php

+++ app/Http/Controllers/LocatarioController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Catalog;
+use App\Models\Conversation;
 use App\Http\Requests\FilteringRequest;
 use App\Models\OfferList;
 use App\Models\Resources\Option;
@@ -12,11 +13,15 @@ class LocatarioController extends Controller {
 
     protected $catalogModel;
     protected $offerListModel;
+    protected $conversationModel;
+
+
 
     public function __construct() {
         $this->middleware('can:isLocatario');
         $this->catalogModel = new Catalog;
         $this->offerListModel = new OfferList;
+        $this->conversationModel = new Conversation;
     }
 
     public function showFilteredCatalog(FilteringRequest $request) {
@@ -32,12 +37,25 @@ class LocatarioController extends Controller {
     }
 
     public function deleteOption($offerId) {
-        $autenticati = Option::where('offerta_id', $offerId)->value('locatario_id');
-        if (Auth::id() == $autenticati) {
-            Option::where('offerta_id', $offerId)->delete();
-            return redirect()->action('LocatarioController@showOptionedList');
-        }
-        return redirect()->action('UserController@index');
+        Option::where('offerta_id', $offerId)->where('locatario_id',Auth::id())->delete();
+        return redirect()->action('LocatarioController@showOptionedList');
+
     }
+    
+    public function optionedOffer($offer_id){
+        $this->offerListModel->insertOption($offer_id, Auth::id());
+        $offer = $this->offerListModel->getOfferByIdOffer($offer_id);
+        $this->conversationModel->createDefaultMessage($offer,Auth::id());
+        return redirect()->action('PublicController@showDetails',$offer_id);             
+    }
+    
+    public function showOptionButton($offer_id){
+         return $this->offerListModel->getOptionByOfferLarioId($offer_id, Auth::id());
+    }
+    
+    public function showButtons($offer_id) {
+        return $this->offerListModel->verifyAssigned($offer_id);
+    }
+   
 
 }
